Refs #392, #342, #317. Fixing float issue in iw_console, adding multiple variations on dependencies of TrackPeople strategies, fixing a bug in ProcModules strategies when filters were not explicitely declared, now defaults to affecting all filters from a module. Preparing for total utility maximisation.
git-svn-id: svn+ssh://marvin.3it.usherbrooke.ca/svn/introlab/johnny0/svn/trunk/ros-pkgs/hbba@3877 23a761d4-b60a-4777-9503-f06a07968eac

+++ iw/src/iw_console.py
@@ -87,11 +87,11 @@ class Shell:
                     continue
             elif (state == 2):
                 state = 3 # AddTypeInt
-                desire.intensity = float(t)
+                desire.intensity = int(t)
                 continue
             elif (state == 3):
                 state = 4 # AddTypeIntParams 
-                desire.utility = float(t)
+                desire.utility = int(t)
                 continue
             elif (state == 4):
                 desire.params += t
